refactor(skills): decompose SkillsModule
PKT-1126
